Made into custom hooks
